Moved rest controllers into a rest subdirectory
