+++ src/controllers/index.js
@@ -15,8 +15,8 @@ const Adapters = require('../adapters')
 const adapters = new Adapters()
 
 // Load the REST API Controllers.
-const PostEntry = require('./post-entry')
-const GetAll = require('./get-all')
+const PostEntry = require('./rest/post-entry')
+const GetAll = require('./rest/get-all')
 
 // Top-level function for this library.
 // Start the various Controllers and attach them to the app.
@@ -27,40 +27,6 @@ async function attachControllers (app) {
   // Start the P2WDB and attach the validation event handler/controller to
   // the add-entry Use Case.
   await attachValidationController()
-}
-
-// Start the P2WDB and its downstream depenencies (IPFS, ipfs-coord, OrbitDB).
-// Also attach the post-validation, peer-replication event handler (controller)
-// to the Add-Entry Use Case.
-async function attachValidationController () {
-  try {
-    // Start the P2WDB.
-    await adapters.p2wdb.start()
-
-    // Trigger the addPeerEntry() use-case after a replication-validation event.
-    adapters.p2wdb.orbit.validationEvent.on(
-      'ValidationSucceeded',
-      async function (data) {
-        try {
-          // console.log(
-          //   'ValidationSucceeded event triggering addPeerEntry() with this data: ',
-          //   data
-          // )
-
-          await useCases.addEntry.addPeerEntry(data)
-        } catch (err) {
-          console.error(
-            'Error trying to process peer data with addPeerEntry(): ',
-            err
-          )
-          // Do not throw an error. This is a top-level function.
-        }
-      }
-    )
-  } catch (err) {
-    console.error('Error in controllers/index.js/startP2wdb()')
-    throw err
-  }
 }
 
 function attachRESTControllers (app) {
@@ -100,6 +66,40 @@ function attachRESTControllers (app) {
 
   // Attach the Controller routes to the Koa app.
   app.use(router.routes()).use(router.allowedMethods())
+}
+
+// Start the P2WDB and its downstream depenencies (IPFS, ipfs-coord, OrbitDB).
+// Also attach the post-validation, peer-replication event handler (controller)
+// to the Add-Entry Use Case.
+async function attachValidationController () {
+  try {
+    // Start the P2WDB.
+    await adapters.p2wdb.start()
+
+    // Trigger the addPeerEntry() use-case after a replication-validation event.
+    adapters.p2wdb.orbit.validationEvent.on(
+      'ValidationSucceeded',
+      async function (data) {
+        try {
+          // console.log(
+          //   'ValidationSucceeded event triggering addPeerEntry() with this data: ',
+          //   data
+          // )
+
+          await useCases.addEntry.addPeerEntry(data)
+        } catch (err) {
+          console.error(
+            'Error trying to process peer data with addPeerEntry(): ',
+            err
+          )
+          // Do not throw an error. This is a top-level function.
+        }
+      }
+    )
+  } catch (err) {
+    console.error('Error in controllers/index.js/startP2wdb()')
+    throw err
+  }
 }
 
 module.exports = { attachControllers }
