+++ src/controllers/rest/get-all.js
