+++ src/controllers/rest/post-entry.js
